+++ Ppap/app/src/main/java/com/jeongg/ppap/data/dto/ScrapListDTO.kt
@@ -1,0 +1,9 @@
+package com.jeongg.ppap.data.dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ScrapListDTO(
+    val curSubscribeId: Long,
+    val scraps: List<ScrapDTO>
+)
