+++ Ppap/app/src/main/java/com/jeongg/ppap/data/dto/SubscribeCreateRequestDTO.kt
@@ -1,4 +1,4 @@
-package com.jeongg.ppap.data.subscribe.dto
+package com.jeongg.ppap.data.dto
 
 import kotlinx.serialization.Serializable
