+++ Ppap/app/src/main/java/com/jeongg/ppap/data/dto/SubscribeUpdateResponseDTO.kt
@@ -1,0 +1,11 @@
+package com.jeongg.ppap.data.dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SubscribeUpdateResponseDTO(
+    val subscribeId: Long = 0,
+    val title: String = "",
+    val noticeLink: String? = null,
+    val isActive: Boolean = false
+)
