+++ Ppap/app/src/main/java/com/jeongg/ppap/data/dto/SubscribeWithContentDTO.kt
@@ -1,0 +1,12 @@
+package com.jeongg.ppap.data.dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SubscribeWithContentDTO(
+    val subscribeId: Long = 0,
+    val title: String = "",
+    val noticeLink: String? = null,
+    val rssLink: String = "",
+    val isActive: Boolean = false
+)
