--- Ppap/app/src/main/java/com/jeongg/ppap/data/subscribe/dto/SubscribeGetResponseDTO.kt
@@ -1,10 +1,0 @@
-package com.jeongg.ppap.data.subscribe.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SubscribeGetResponseDTO(
-    val subscribeId: Long,
-    val title: String,
-    val isActive: Boolean
-)
